Support removing trends widgets without cogwheel (#1)
Some users are reporting a version of the trends widget without the
cogwheel link. We can try to identify this version (also the other
version, but let’s keep the old code for that) using the “trending
with…” links under some trends, which, if they exist, identify the
source via some URL parameters.