Remove trends a bit more often
With the navigation listener added in the previous commit, most of the
time a call to removeTrends() immediately when receiving the message was
enough to remove the trends, but occasionally they would remain
(presumably Twitter had not finished adding them by the time the message
was dispatched), so I added a setTimeout(removeTrends, 3000) similar to
the one for the page load. However, three seconds is fairly long to wait
– most of the time, Twitter does’t need that long to add the trends. So
add two more timeouts, for 100 and 500 milliseconds, and while we’re at
it, also add those to the initial page load via a helper function. (This
also resolves the issue that the trends would not be immediately removed
after installing the extension and going back to an already-loaded
Twitter tab, which was potentially confusing.)