Identify another version of the Trends widget
And also return as soon as we’ve removed the first widget, to reduce the
performance impact of the extension. I’m assuming that there can only
ever be one Trends widget on the page, and if the extension has any
false positives (removing something that isn’t actually a Trends
widget), this should make that easier to notice, since the real Trends
widget will now still be around.